# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: llama_shim API
  version: 0.1.0
  summary: Minimal OpenAI-compatible surface owned by llama_shim
  description: |
    Versioned API contract for the routes currently implemented or explicitly
    owned by `llama_shim`.

    The spec is intentionally narrow:
    - only routes that exist in the current server are documented
    - operation status is marked with `x-shim-status`
    - `implemented` means the shim owns the route contract
    - `partial` means the shim owns only part of the route behavior or request surface
    - `proxy` means the route is validated and then forwarded to the upstream backend

    Notes:
    - `/v1/responses` is local-first for the supported subset and may still proxy
      unsupported features upstream depending on `responses.mode`
    - the shim may enforce optional ingress bearer auth, request rate limiting,
      request/upload/runtime quotas, and a shim-owned `/metrics` endpoint;
      these are operational controls owned by the shim, not OpenAI parity
      claims
    - item payloads are intentionally described as typed open objects because the
      shim preserves normalized raw item bodies across several Responses and
      Conversations flows
servers:
  - url: http://127.0.0.1:8080
...
paths:
...
  /debug/capabilities:
    get:
      tags: [ops]
      summary: Shim-owned capability manifest
      x-shim-status: implemented
      description: |
        Returns a shim-owned capability manifest describing the current
        process configuration, locally owned surfaces, local-vs-proxy tool
        routing classes, and dependency probe state.

        Notes:
        - this route is shim-owned operator/debug surface, not an OpenAI API
          parity claim
        - unlike `/healthz` and `/readyz`, this route shares normal shim
          ingress auth and request rate limiting
        - the route returns `200` even when one or more dependencies are
          degraded; inspect top-level `ready` and `probes.*`
      responses:
        "200":
          description: Capability manifest for the current shim process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShimCapabilitiesResponse"
        "401":
          description: Missing or invalid shim ingress bearer token when auth is enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  message: missing or invalid bearer token
                  type: authentication_error
                  param: null
                  code: null
        "429":
          description: Shim ingress request rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  message: request rate limit exceeded
                  type: rate_limit_error
                  param: null
                  code: rate_limit_exceeded
        "405":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/responses/input_tokens:
    post:
      tags: [responses]
      summary: Count input tokens for a response request
      x-shim-status: partial
      description: |
        Returns `object: "response.input_tokens"` with an `input_tokens` count.

        Shim behavior:
        - for shim-local stateful requests, the shim expands local
          `previous_response_id`, `conversation`, and synthetic compaction items
          into the effective local context snapshot and returns a deterministic
          token estimate for that rendered snapshot
        - this local count is a compatibility estimate, not backend-native
          tokenization
        - requests that do not depend on shim-local state may be proxied
          upstream unchanged for exact provider-native counting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponseInputTokensRequest"
            examples:
              minimal:
                value:
                  model: test-model
                  input: Count this request
              previous_response:
                value:
                  model: test-model
                  previous_response_id: resp_prev_123
                  input: What was my code?
      responses:
        "200":
          description: Token count result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseInputTokensResource"
              example:
                object: response.input_tokens
                input_tokens: 11
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
        "504":
          $ref: "#/components/responses/ErrorResponse"
  /v1/responses/compact:
    post:
      tags: [responses]
      summary: Compact a response context window
      x-shim-status: partial
      description: |
        Returns a `response.compaction` resource.

        Shim behavior:
        - local compaction returns a canonical next input window that can be
          passed into the next local `/v1/responses` request as-is
        - the default heuristic backend returns a shim-owned opaque
          `compaction` item; the `model_assisted_text` backend may include a
          bounded retained tail followed by the compaction item
        - on the next local request, the shim expands that opaque blob into a
          synthetic summary message before generation
        - this is not byte-for-byte equivalent to OpenAI encrypted compaction
          state; it is a compatible local continuation mechanism for the
          shim-supported subset
        - requests that do not depend on shim-local state may be proxied
          upstream unchanged
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponseCompactionRequest"
            examples:
              minimal:
                value:
                  model: test-model
                  input:
                    - type: message
                      role: user
                      content: Remember that the launch code is 777.
                    - type: message
                      role: assistant
                      content:
                        - type: output_text
                          text: I will remember the launch code.
      responses:
        "200":
          description: Shim-owned compaction resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseCompactionResource"
              example:
                id: resp_123
                object: response.compaction
                created_at: 1741900000
                output:
                  - type: message
                    role: user
                    content:
                      - type: input_text
                        text: Remember that the launch code is 777.
                  - id: cmp_123
                    type: compaction
                    encrypted_content: llama_shim.compaction.v1:ZXhhbXBsZQ
                usage:
                  input_tokens: 42
                  input_tokens_details:
                    cached_tokens: 0
                  output_tokens: 20
                  output_tokens_details:
                    reasoning_tokens: 0
                  total_tokens: 62
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
        "504":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/responses/{id}/cancel:
    post:
      tags: [responses]
      summary: Cancel a background response
      x-shim-status: partial
      description: |
        Cancels a background response.

        Behavior:
        - locally stored non-background responses are rejected with
          `invalid_request_error`
        - locally stored background responses are proxied upstream and the local
          shadow copy is refreshed from the upstream response object
        - missing local ids are forwarded to upstream `/v1/responses/{id}/cancel`
      parameters:
        - $ref: "#/components/parameters/ResponseID"
      responses:
        "200":
          description: Updated response object after cancellation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
        "504":
          $ref: "#/components/responses/ErrorResponse"
  /v1/responses/{id}/input_items:
    get:
      tags: [responses]
      summary: List input items for a stored response
      x-shim-status: partial
      description: |
        Returns the locally stored snapshot of input items used to generate the
        response when present. For shim-owned previous-response and
        conversation-backed turns, this includes the effective stateful context
        assembled for generation, not just the current turn's raw input. If the
        response is not found in shim storage, the request is forwarded to
        upstream `/v1/responses/{id}/input_items`.

        Shim-owned `previous_response_id` lineage reconstruction is bounded by
        the internal `shim.limits.responses_stored_lineage_max_items` runtime
        knob. The knob is not an OpenAI request parameter and does not add a
        public 4xx limit; it only bounds how much local stored history the shim
        reconstructs for context and legacy input-item snapshots.

        Supported shim-owned query surface for locally stored responses:
        - `after`, `limit`, and `order` paginate the locally stored normalized
          input-item snapshot
        - `include` is accepted as a compatibility no-op
      parameters:
        - $ref: "#/components/parameters/ResponseID"
        - $ref: "#/components/parameters/ListAfter"
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListOrder"
        - $ref: "#/components/parameters/ResponseInclude"
      responses:
        "200":
          description: Input items stored for the response generation snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
              example:
                object: list
                data:
                  - id: msg_user_1
                    type: message
                    role: user
                    content:
                      - type: input_text
                        text: Remember code 123
                  - id: msg_assistant_1
                    type: message
                    role: assistant
                    content:
                      - type: output_text
                        text: OK
                  - id: msg_user_2
                    type: message
                    role: user
                    content:
                      - type: input_text
                        text: Tell me more
                first_id: msg_user_1
                last_id: msg_user_2
                has_more: false
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /v1/conversations:
    post:
      tags: [conversations]
      summary: Create a conversation
      x-shim-status: partial
      description: |
        Creates a shim-owned conversation from initial items.

        Compared with the official OpenAI Conversations API surface:
        - empty-body create is supported
        - the shim supports the documented `items` seed flow and lightweight
          string `metadata`
        - the returned conversation resource matches the official top-level
          shape more closely
        - conversation items are accessed through the dedicated item endpoints
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationRequest"
            example:
              metadata:
                topic: demo
              items:
                - type: message
                  role: system
                  content: You are a test assistant.
                - type: message
                  role: user
                  content: Remember code 777. Reply OK.
      responses:
        "200":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
              example:
                id: conv_123
                object: conversation
                created_at: 1741900000
                metadata:
                  topic: demo
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /v1/conversations/{id}:
    get:
      tags: [conversations]
      summary: Get a conversation
      x-shim-status: partial
      description: |
        Returns the conversation resource for the given id.

        Compared with the official OpenAI Conversations API surface:
        - the resource shape follows the official top-level conversation object
        - conversation items are read through the dedicated item endpoints
      parameters:
        - $ref: "#/components/parameters/ConversationID"
      responses:
        "200":
          description: Conversation resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
              example:
                id: conv_123
                object: conversation
                created_at: 1741900000
                metadata:
                  topic: demo
        "404":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/conversations/{id}/items/{item_id}:
    get:
      tags: [conversations]
      summary: Get a conversation item
      x-shim-status: partial
      parameters:
        - $ref: "#/components/parameters/ConversationID"
        - $ref: "#/components/parameters/ConversationItemID"
        - name: include
          in: query
          description: |
            Additional item fields to include in the response. The shim validates
            the documented include values and accepts them as a compatibility
            no-op because stored item payloads are already returned in full.
          schema:
            type: array
            items:
              type: string
              enum:
                - computer_call_output.output.image_url
                - code_interpreter_call.outputs
                - file_search_call.results
                - web_search_call.action.sources
      responses:
        "200":
          description: Conversation item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
              example:
                id: item_3
                type: message
                role: user
                content:
                  - type: input_text
                    text: Remember city=Paris. Reply OK.
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [conversations]
      summary: Delete a conversation item
      x-shim-status: partial
      description: Delete an item from the conversation and return the conversation resource.
      parameters:
        - $ref: "#/components/parameters/ConversationID"
        - $ref: "#/components/parameters/ConversationItemID"
      responses:
        "200":
          description: Conversation resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
              example:
                id: conv_123
                object: conversation
                created_at: 1741900000
                metadata:
                  topic: demo
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/files/{file_id}:
    get:
      tags: [retrieval]
      summary: Retrieve a locally stored file
      x-shim-status: partial
      parameters:
        - $ref: "#/components/parameters/FileID"
      responses:
        "200":
          description: Stored file resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileObject"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [retrieval]
      summary: Delete a locally stored file
      x-shim-status: partial
      description: |
        Deletes a shim-owned local file and removes it from any locally managed
        vector stores attached to that file.
      parameters:
        - $ref: "#/components/parameters/FileID"
      responses:
        "200":
          description: File deletion marker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDeletion"
        "404":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  parameters:
    ResponseID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Response identifier
...
  schemas:
...
    ShimCapabilitiesResponse:
      type: object
      required: [object, ready, surfaces, runtime, tools, probes]
      properties:
        object:
          type: string
          enum: [shim.capabilities]
        ready:
          type: boolean
        surfaces:
          $ref: "#/components/schemas/ShimCapabilitySurfaceSet"
        runtime:
          $ref: "#/components/schemas/ShimCapabilityRuntime"
        tools:
          $ref: "#/components/schemas/ShimCapabilityToolSet"
        probes:
          $ref: "#/components/schemas/ShimCapabilityProbeSet"
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
...